requirements.txt
python-socketio==5.4.0
aiohttp[speedups]
